part 2 of that rename
