+++ shepherd_core/shepherd_core/vsource/virtual_harvester_simulation.py
@@ -13,8 +13,8 @@ from pathlib import Path
 from tqdm import tqdm
 
 from shepherd_core.data_models.base.calibration import CalibrationHarvester
-from shepherd_core.data_models.content.virtual_harvester import HarvesterPRUConfig
-from shepherd_core.data_models.content.virtual_harvester import VirtualHarvesterConfig
+from shepherd_core.data_models.content.virtual_harvester_config import HarvesterPRUConfig
+from shepherd_core.data_models.content.virtual_harvester_config import VirtualHarvesterConfig
 from shepherd_core.reader import Reader
 from shepherd_core.writer import Writer
