+++ shepherd_core/shepherd_core/vsource/virtual_source_model.py
@@ -12,9 +12,9 @@ NOTE: DO NOT OPTIMIZE -> stay close to original code-base
 
 from shepherd_core.data_models.base.calibration import CalibrationEmulator
 from shepherd_core.data_models.content.energy_environment import EnergyDType
-from shepherd_core.data_models.content.virtual_harvester import HarvesterPRUConfig
-from shepherd_core.data_models.content.virtual_source import ConverterPRUConfig
-from shepherd_core.data_models.content.virtual_source import VirtualSourceConfig
+from shepherd_core.data_models.content.virtual_harvester_config import HarvesterPRUConfig
+from shepherd_core.data_models.content.virtual_source_config import ConverterPRUConfig
+from shepherd_core.data_models.content.virtual_source_config import VirtualSourceConfig
 from shepherd_core.data_models.content.virtual_storage_config import StoragePRUConfig
 
 from .virtual_converter_model import PruCalibration
